sysdep rework

+++ make-tinyssh.sh
@@ -132,14 +132,19 @@ mkdir -p "${work}"
 cp -pr sysdep/* "${work}"
 (
   cd "${work}"
-  echo 'int main(void) { return 0; }' > main.c
   sh list | (
     while read target source
     do
       [ -f "${include}/${target}" ] && continue
       rm -f "${source}" "${target}.tmp" 
-      ${compiler} -O0 -o "${source}" "${source}.c" main.c ${libs} || continue
-      cp "${source}.out" "${include}/${target}"
+      ${compiler} -O0 -o "${source}" "${source}.c" ${libs} || continue
+      "./${source}" > "${target}.tmp" 2>/dev/null || continue
+      if [ -f "${source}.out" ]; then
+        cp "${source}.out" "${include}/${target}"
+      else
+        #runtime
+        cp "${target}.tmp" "${include}/${target}"
+      fi
       log2 "${target} ${source}"
     done
   )
